feat(citation): add `Citation`
Citations sometimes must be changed due to a new one coming in, for
example, due to disambiguating names.

Previously, this was not done, as a simple string was returned by `cite`.

Now, `cite` returns a `Citation` object, which as a `value` of the citation
dynamically updated by a processor.

Additionally temporarily removes `citeInText` functionality.